[CHORE] refactor tests from the root folder to the folder tests
